Add cleanup script for failed builds
This does not reset any of the stages (for that use reset.sh), it only
unmounts all of the chroot mountpoints from every stage, which is
harmless if they don't exist.

It then tries to unmap any images that no longer have any active loop
device attached. That step can sometimes fail, but it should be
harmless if it does.

We need to do this because a failed build can leave some of those
things still mounted, which complicates running another build right
afterward and can lead to odd build problems.