Add git commit of builder/open.hd repo to images
Only the Open.HD repo version is used in the image name itself, as
putting 2 different versions in the image name would make it long and
confusing. Both versions are still added to files in the root
of the image itself, so it is always possible to trace an image
back to the source code used to build it even if it gets renamed.

When the current commit has an actual tag like v0.1.0, then the tag
itself will be used as the version rather than a git commit hash. This
enables sane versioning just by tagging the builder repo.

If there have been commits added to the repo since the last tag, then
the git describe --tags format will be used: <tag>-<numcommits>-<hash>,
showing the *last* tag, the number of commits since that tag, and the
*current* commit.

If no tag is present in the repo yet, then the hash of the current
commit will be used.

To keep everything properly version controlled we can either make the
Open.HD repo a submodule of the builder repo (which then means that
a particular commit in the builder repo is linked to a particular
commit in the Open.HD repo), or just add a commit hash for a particular
commit to the CONFIG file in the root of the builder, and have
the scripts checkout that commit when building. The latter is going to
be easier for people who aren't familiar with git.