Mount root filesystem rw
Why:

Mounting root read-only causes problems with anything that needs to
save settings or alter the filesystem at runtime, like QOpenHD. It
doesn't actually solve the problem it's attempting to solve either, at
least in the case of the root filesystem. For the /boot filesystem it
might be wise to leave it read-only for now since /boot is a fat32
filesystem with no journaling.

Details:

Mounting read-only is an attempt to prevent filesystem corruption on
the pi. However, the usual cause of that corruption is not something
you can fix or address by mounting read-only, it's caused by sudden
power loss, undervoltage, overvoltage, and poor quality SD cards. In
other words, it's the card causing the corruption, not writes to the
filesystem.

SD cards run maintenance tasks on their own flash at arbitrary times
for wear leveling and to deal with read disturb[1]. This is one of the
real reasons the Pi has issues with SD card corruption. Sudden power
loss can interrupt that maintenance, and undervoltage can cause cheap
SD card controllers to malfunction.

You can't prevent any of that by making the filesystem read-only,
the card has no idea whether the filesystem is mounted that way or
not and will carry out maintenance tasks whenever it wants.

The only way to prevent that from causing filesystem corruption is to
use SD cards that can handle sudden power loss, like industrial cards
that are designed for it.

A better solution to issues that do affect the filesystem layer would
be to significantly shorten the journal flush time or ensure that the
sync() call or sync command are used where necessary to force the
filesystem to write pending changes to the flash.

[1] https://en.wikipedia.org/wiki/Flash_memory#Read_disturb