Add support for apt-cacher-ng in builder
With this change, if you install apt-cacher-ng on your build machine
and enable it to cache raspbian packages (if it doesn't by default),
the first time you run a build the packages being installed in stage3
will be cached on your build machine. Then the next time you run a new
build, it will get those packages from apt-cacher-ng almost instantly
instead of having to download them again. The speedup is massive, 10
minutes vs over an hour depending on the connection speed