Copy openhd-settings-1.txt to 2/3/4 during build
We don't really need to have 4 separate copies of the same file
in git, it creates opportunities for one of them to become outdated.

Instead we can copy the real file 3 times during a build so that the
final image always ends up with 4 copies of the exact same settings
file.