Add temporary special case for CM3+ images
This adds an argument to build.sh to allow building images for other
pi models that do not work with the current base image, like the CM3+.

Once we know the newer image works for the other pi models as well we
can simply update it and remove the special casing.

Pi4 support will be added in a similar way, a temporary special case
until the newer buster image is well tested on pi zero and pi3.