Add buster image for pi4/cm3+, simplify stage 2
This moves the CM3+ image to a more general buster image, which also
gives us pi4 support.

We still build our own kernels but instead of patching manually we
just use our own tree which has branches for each kernel version and
patches already applied. This makes it far easier to update the kernel
or pull in specific bugfixes from the Raspbian developers or mainline
linux if we need to.

The linux build is done in a separate directory for each kernel branch
which prevents accidentally building a stretch image with a 4.19
kernel or a buster image with a 4.14 kernel. This is temporary as
buster is going to be the default image soon.