Enable TTYs by name instead of copying unit files
This was expecting systemd to proactively start all these TTYs even
if the user had not switched to them in the local console. That was
true by default on stretch, but is not true on buster.

We use the TTYs as a form of parallelism during boot, so they all
need to start anyway regardless of whether the user is even looking
at the local console.