Enable the fKMS driver stack on buster
We still use the Broadcom driver on stretch, and we're going to leave
that alone since it works fine.

It's also likely there are bugfixes being applied to the Mesa/DRM
packages and 4.19 kernel in buster, but not the Mesa/DRM packages
and 4.14 kernel in stretch, which are not going to be maintained
upstream much longer, if they're even still being maintained.