Set FLIR camera pipelines to 30fps
The camera itself is sending a stream close to 30fps because the Lepton
module does frame duplication from 9hz to around 27hz. So when we
ask for 10fps we're essentially throwing away 1/3rd of the real thermal
frames and getting something that barely works.

Requesting a 30fps stream gets us the full 9hz thermal images.

This was tested by Thomas and seems to work significantly better than
the previous line.