Don't advertise gateway/dns to hotspot DHCP client
The dnsmasq service was telling hotspot clients that the ground
station was a router capable of being a default gateway, and to route
DNS requests to it.

Advertising a default gateway and DNS causes android phones to route
traffic to the ground station even when they could use LTE normally,
which makes it impossible to use anything requiring internet access
while connected to the openhd hotspot.