Enable atomic KMS API in QOpenHD and OpenHDBoot
This fixes an issue where the screen will go blank when certain
joystick/TX controllers are connected to the ground station. It's
a ridiculous bug somewhere in the Qt+fKMS display stack in Buster.

Even switching to atomic KMS doesn't fix it completely, there is
still a brief blank flash for a fraction of a second, only for the
first Qt application to run, so in our case that means the boot
screen.

We didn't enable the atomic API in the past because another Qt bug
caused it to completely fail to work on the Pi3b+ on buster when
the official pi touchscreen was connected (HDMI worked fine in that
case). As of Qt 5.14 or 5.15 that one is fixed so we can enable it
now, but we still need to figure out why the screen flash happens.