Prevent OpenHDBoot, QOpenHD from changing KMS mode
With certain HDMI devices, the Pi firmware and the Linux kernel don't
have the same view of what the preferred screen resolution/mode should
be.

As a result, the pi firmware picks one resolution/mode, but then the
boot screen and QOpenHD are told by the linux kernel that a different
resolution is preferred, so they switch to it.

This severely breaks things, resulting in missing video and other
problems, but only affecting devices like the Headplay goggles.

Only affects Buster because Stretch is not using the KMS display
stack.