Use packages for linux/QOpenHD/router/microservice
This removes the entire stage 2 of the builder and uses a package
for the kernel. It removes the microservice and router build steps and
uses packages for them, uses packages for qt and qopenhd instead
of the simple tarballs we used before.

The openhd repo is removed at the end of the installation steps to
ensure the image remains the same after being built, as upgrades are
not yet tested and will require some more attention.

This slashes at least an hour off the image build time, probably 2
hours on a slow machine such as a cloud VM.