Update repo location for 2.0 repo
The repos for 2.0 and 2.1 are separate to ensure that 2.0 users are
only receiving bugfixes and not new stuff that could affect their
existing setup.

Upgrades from 2.0 to 2.1 will be done by the microservice code, it
will replace the file in /etc/apt/sources.list.d/ and then trigger
update && upgrade of the openhd package.