Remove version constraint on openhd package
There are no versions yet, and we'll need to consider how best to
name the tags to be compatible with the apt package version scheme,
while still allowing us to do things like release betas. The previous
scheme "2.0.0rc1.2" results in apt thinking 2.0.0rc1.2 is > 2.0.0. I
have seen some projects do 2.0.9999 or something similar to make clear
it's almost 2.1.0 but not quite.