Use target profiles
This allows for all of the details for a particular target to be
kept separate from the rest of the builder, making it easy to expand
or change the target.

For the most part things should not have different behavior on
different targets, the only exception will be platform-specific
packages that must be held back with apt or differences in the way
the base images are organized (some don't have a boot partition).