Only disable dir_index if base image is 32-bit
When running on a 64-bit host with qemu inside a 32-bit base image,
there are glibc quirks that can cause strange errors that will not
make sense and will send developers off chasing things that aren't
really happening, such as error messages about missing files or
directories even though they're present.

The workaround in this case is to disable dir_index on ext4, because
it alters the way some syscalls are used in glibc and avoids the
problem. This is something that should really be fixed in qemu and
at some point it will be if it isn't already but we can't depend
on that and can't wait.