Add NanoPi Neo2/3 targets
These are not ready for full use because the base images do not have
a boot or conf partition, however the only practical effect of that
is that the settings files don't show up on a windows PC the way
the other images do. The system will still work until we adjust the
base images as everything is fully automatic anyway.

We will either need to add a /conf partition manually in the images
from the download host (just re-upload them), make the builder do it
with scripting (not ideal because the partition would be after root,
making root impossible to resize), or use something other than this
builder